refactor: replace inline resource fetch logic with `downloadProxyResource` utility for cleaner and reusable code across components
Hook up a view/download file menu item in FileManager.vue
