+++ src/lib/components/chat/Chat.svelte
@@ -142,6 +142,7 @@ $: {
 // Auto open/close of Private AI sidekick is now handled via appHooks 'model.changed' in onMount.
 import { appHooks, type PrivateAiExtras } from '$lib/utils/hooks';
 import { apiFetch } from '$lib/apis/private-ai/fetchClients';
+import { downloadProxyResource } from '$lib/utils/privateAi';
 
 let __unhookModelChanged: (() => void) | undefined;
 let __unhookChatSubmit: (() => void) | undefined;
@@ -227,151 +228,13 @@ async function handleGotoClick(e: MouseEvent) {
 			return;
 		}
 
-		// Allowlist for common development hostnames
-		const devHostnames = new Set(['localhost', '127.0.0.1', '0.0.0.0', '::1']);
-		const pageIsLocal = devHostnames.has(window.location.hostname);
-		const targetIsLocal = devHostnames.has(targetUrl.hostname);
-		const serviceIsLocal = devHostnames.has(serviceBase.hostname);
-
-		// Accept only when the target URL is fetched by the configured service:
-		// 1) In production: require exact same origin as configured service (serviceBase.origin),
-		//    AND the target path must be equal to or under the configured service base path.
-		// 2) In local dev: allow local-to-local flows (page local && target local) or when the configured
-		//    service itself is local and the target is local.
-		const originAllowed =
-			targetUrl.origin === serviceBase.origin ||
-			(pageIsLocal && targetIsLocal) ||
-			(serviceIsLocal && targetIsLocal);
-
-		if (!originAllowed) {
-			console.warn('Link origin not allowed by policy', {
-				target: targetUrl.toString(),
-				serviceBase: serviceBase.toString(),
-				pageHostname: window.location.hostname
-			});
-			toast.error('Link refused: origin mismatch for linked resource.');
-			return;
-		}
-
-		// Enforce that the clicked path is under the configured service path (if a service path provided).
-		// This prevents accepting unrelated paths on the same origin.
-		const servicePath = (serviceBase.pathname || '').replace(/\/$/, '');
-		if (servicePath) {
-			const p = targetUrl.pathname || '';
-			// Allow when the target path exactly equals servicePath or is under it (component boundary)
-			if (!(p === servicePath || p.startsWith(servicePath + '/'))) {
-				console.warn('Link path not under configured service path', {
-					targetPath: p,
-					servicePath
-				});
-				toast.error('Link refused.');
-				return;
-			}
-		}
-
-		// OK: targetUrl is permitted to be fetched via the configured service.
-		// Use apiFetch.raw() to get the actual Response (ofetch exposes .raw()).
-		let resp: Response | null = null;
+		// Delegate the allowlist/origin/path validation and download/open behavior to the shared utility.
+		// The util will validate the URL against backendConfig and only proceed if allowed.
 		try {
-			resp = await apiFetch.raw(targetUrl.toString());
+			await downloadProxyResource(targetUrl, backendConfig);
 		} catch (err) {
-			console.error('apiFetch.raw failed for goto URL', err);
+			console.error('Failed to download/open resource', err);
 			toast.error('Failed to retrieve resource');
-			return;
-		}
-
-		// Expect a Response containing the downloadable blob.
-		try {
-			if (!resp || !resp.ok) {
-				console.warn('Non-OK response from proxy download', resp);
-				toast.error('Resource unavailable');
-				return;
-			}
-
-			const blob = resp._data; // ofetch provides this property
-			if (!(blob instanceof Blob) || blob.size === 0) {
-				console.warn('Invalid or empty Blob returned for download', resp);
-				toast.error('Resource unavailable');
-				return;
-			}
-
-			// Derive filename from response.url which for this proxy always ends with '/proxy-download'.
-			// The pattern is: /files/{filePath}/proxy-download where filePath ends with the actual filename.
-			let filename = 'download';
-			try {
-				const responseUrl = typeof resp.url === 'string' && resp.url ? resp.url : targetUrl.toString();
-				const u = new URL(responseUrl, window.location.href);
-				const pathname = u.pathname || '';
-				const proxySuffix = '/proxy-download';
-
-				let candidate = '';
-				if (pathname.endsWith(proxySuffix)) {
-					// Strip suffix and take the last segment of the remaining path
-					const filePath = pathname.slice(0, -proxySuffix.length);
-					candidate = filePath.split('/').filter(Boolean).pop() ?? '';
-				} else {
-					// Fallback: use the last segment of the pathname
-					candidate = pathname.split('/').filter(Boolean).pop() ?? '';
-				}
-
-				if (candidate) {
-					try {
-						filename = decodeURIComponent(candidate);
-					} catch {
-						filename = String(candidate);
-					}
-					// Sanitize filename (keep alphanumerics, dot, underscore, hyphen)
-					filename = filename.replace(/[^a-zA-Z0-9.\-_]/g, '_');
-				}
-			} catch (e) {
-				// ignore and use default filename
-				filename = 'download';
-			}
-
-			// Trim filename to reasonable length
-			if (filename.length > 200) {
-				const extMatch = filename.match(/(\.[a-zA-Z0-9]{1,8})$/);
-				const ext = extMatch ? extMatch[1] : '';
-				const namePart = filename.slice(0, 200 - ext.length);
-				filename = `${namePart}${ext}`;
-			}
-
-			// Create an object URL. If it's a PDF, open in a new tab; otherwise trigger a secure download.
-			const objectUrl = URL.createObjectURL(blob);
-
-			const isPdf = (blob.type === 'application/pdf') || /\.pdf$/i.test(filename);
-
-			let opened = false;
-			if (isPdf) {
-				// Attempt to open in a new tab. Because this runs in a click handler, most browsers will allow it.
-				const w = window.open(objectUrl, '_blank', 'noopener,noreferrer');
-				opened = !!w;
-			}
-
-			if (!opened) {
-				// Fallback to download (or non-PDFs)
-				const a = document.createElement('a');
-				a.href = objectUrl;
-				if (!isPdf) a.download = filename;
-				a.rel = 'noopener noreferrer';
-				document.body.appendChild(a);
-				a.click();
-				a.remove();
-			}
-
-			// Revoke the object URL after a short timeout to allow the browser to consume it.
-			setTimeout(() => {
-				try {
-					URL.revokeObjectURL(objectUrl);
-				} catch (e) {
-					/* ignore revoke errors */
-				}
-			}, 60000);
-
-			toast.success($i18n.t(isPdf ? 'Opened in a new tab' : 'Download started'));
-		} catch (err) {
-			console.error('Failed to process downloaded Blob', err);
-			toast.error('Failed to download resource');
 		}
 	} catch (err) {
 		console.error('handleGotoClick error', err);
